@@ -40,9 +40,10 @@ var App = React.createClass({
 var Authentication = {
   statics: {
     willTransitionTo: function (transition) {
+      var nextPath = transition.path;
       if (!auth.loggedIn()) {
-        Login.attemptedTransition = transition;
-        transition.redirect('/login');
+        transition.redirect('/login',{},
+          { 'nextPath' : nextPath });
       }
     }
   }
@@ -64,10 +65,9 @@ var Dashboard = React.createClass({
 });
 
 var Login = React.createClass({
-  mixins: [ Router.Navigation ],
 
-  statics: {
-    attemptedTransition: null
+  contextTypes: {
+    router: React.PropTypes.func.isRequired
   },
 
   getInitialState: function () {
@@ -78,18 +78,18 @@ var Login = React.createClass({
 
   handleSubmit: function (event) {
     event.preventDefault();
+    var { router } = this.context;
+    var nextPath = router.getCurrentQuery().nextPath;
     var email = this.refs.email.getDOMNode().value;
     var pass = this.refs.pass.getDOMNode().value;
     auth.login(email, pass, function (loggedIn) {
       if (!loggedIn)
         return this.setState({ error: true });
 
-      if (Login.attemptedTransition) {
-        var transition = Login.attemptedTransition;
-        Login.attemptedTransition = null;
-        transition.retry();
+      if (nextPath) {
+        router.replaceWith(nextPath);
       } else {
-        this.replaceWith('/about');
+        router.replaceWith('/about');
       }
     }.bind(this));
   },
@@ -168,7 +168,7 @@ function pretendRequest(email, pass, cb) {
     if (email === 'joe@example.com' && pass === 'password1') {
       cb({
         authenticated: true,
-        token: Math.random().toString(36).substring(7),
+        token: Math.random().toString(36).substring(7)
       });
     } else {
       cb({authenticated: false});
