@@ -17,10 +17,13 @@ var App = React.createClass({
 });
 
 var User = React.createClass({
-  mixins: [ Router.State ],
+
+  contextTypes: {
+    router: React.PropTypes.func.isRequired
+  },
 
   render () {
-    var { userId } = this.getParams();
+    var { userId } = this.context.router.getCurrentParams();
     return (
       <div className="User">
         <h1>User id: {userId}</h1>
@@ -36,10 +39,13 @@ var User = React.createClass({
 
 
 var Task = React.createClass({
-  mixins: [ Router.State ],
+
+  contextTypes: {
+    router: React.PropTypes.func.isRequired
+  },
 
   render () {
-    var { userId, taskId } = this.getParams();
+    var { userId, taskId } = this.context.router.getCurrentParams();
     return (
       <div className="Task">
         <h2>User id: {userId}</h2>
