@@ -1,16 +1,28 @@
 var assign = require('react/lib/Object.assign');
 var ReactPropTypes = require('react').PropTypes;
+var Route = require('./Route');
 
-var PropTypes = assign({
+var PropTypes = assign({}, ReactPropTypes, {
 
   /**
-   * Requires that the value of a prop be falsy.
+   * Indicates that a prop should be falsy.
    */
-  falsy: function (props, propName, componentName) {
+  falsy(props, propName, componentName) {
     if (props[propName])
-      return new Error('<' + componentName + '> may not have a "' + propName + '" prop');
-  }
+      return new Error(`<${componentName}> may not have a "${propName}" prop`);
+  },
 
-}, ReactPropTypes);
+  /**
+   * Indicates that a prop should be a Route object.
+   */
+  route: ReactPropTypes.instanceOf(Route),
+
+  /**
+   * Indicates that a prop should be a Router object.
+   */
+  //router: ReactPropTypes.instanceOf(Router) // TODO
+  router: ReactPropTypes.func
+
+});
 
 module.exports = PropTypes;
