@@ -1,4 +1,16 @@
+var warning = require('react/lib/warning');
 var PropTypes = require('./PropTypes');
+
+function deprecatedMethod(routerMethodName, fn) {
+  return function () {
+    warning(
+      false,
+      `Router.State is deprecated. Please use this.context.router.${routerMethodName}() instead`
+    );
+
+    return fn.apply(this, arguments);
+  };
+}
 
 /**
  * A mixin for components that need to know the path, routes, URL
@@ -8,7 +20,7 @@ var PropTypes = require('./PropTypes');
  *
  *   var AboutLink = React.createClass({
  *     mixins: [ Router.State ],
- *     render: function () {
+ *     render() {
  *       var className = this.props.className;
  *   
  *       if (this.isActive('about'))
@@ -21,56 +33,51 @@ var PropTypes = require('./PropTypes');
 var State = {
 
   contextTypes: {
-    getCurrentPath: PropTypes.func.isRequired,
-    getCurrentRoutes: PropTypes.func.isRequired,
-    getCurrentPathname: PropTypes.func.isRequired,
-    getCurrentParams: PropTypes.func.isRequired,
-    getCurrentQuery: PropTypes.func.isRequired,
-    isActive: PropTypes.func.isRequired
+    router: PropTypes.router.isRequired
   },
 
   /**
    * Returns the current URL path.
    */
-  getPath: function () {
-    return this.context.getCurrentPath();
-  },
-
-  /**
-   * Returns an array of the routes that are currently active.
-   */
-  getRoutes: function () {
-    return this.context.getCurrentRoutes();
-  },
+  getPath: deprecatedMethod('getCurrentPath', function () {
+    return this.context.router.getCurrentPath();
+  }),
 
   /**
    * Returns the current URL path without the query string.
    */
-  getPathname: function () {
-    return this.context.getCurrentPathname();
-  },
+  getPathname: deprecatedMethod('getCurrentPathname', function () {
+    return this.context.router.getCurrentPathname();
+  }),
 
   /**
    * Returns an object of the URL params that are currently active.
    */
-  getParams: function () {
-    return this.context.getCurrentParams();
-  },
+  getParams: deprecatedMethod('getCurrentParams', function () {
+    return this.context.router.getCurrentParams();
+  }),
 
   /**
    * Returns an object of the query params that are currently active.
    */
-  getQuery: function () {
-    return this.context.getCurrentQuery();
-  },
+  getQuery: deprecatedMethod('getCurrentQuery', function () {
+    return this.context.router.getCurrentQuery();
+  }),
+
+  /**
+   * Returns an array of the routes that are currently active.
+   */
+  getRoutes: deprecatedMethod('getCurrentRoutes', function () {
+    return this.context.router.getCurrentRoutes();
+  }),
 
   /**
    * A helper method to determine if a given route, params, and query
    * are active.
    */
-  isActive: function (to, params, query) {
-    return this.context.isActive(to, params, query);
-  }
+  isActive: deprecatedMethod('isActive', function (to, params, query) {
+    return this.context.router.isActive(to, params, query);
+  })
 
 };
 
