@@ -1,22 +1,26 @@
 var LocationActions = require('../actions/LocationActions');
 var History = require('../History');
 
-/**
- * Returns the current URL path from the `hash` portion of the URL, including
- * query string.
- */
-function getHashPath() {
-  return decodeURI(
-    // We can't use window.location.hash here because it's not
-    // consistent across browsers - Firefox will pre-decode it!
-    window.location.href.split('#')[1] || ''
-  );
-}
-
+var _listeners = [];
+var _isListening = false;
 var _actionType;
 
+function notifyChange(type) {
+  if (type === LocationActions.PUSH)
+    History.length += 1;
+
+  var change = {
+    path: HashLocation.getCurrentPath(),
+    type: type
+  };
+
+  _listeners.forEach(function (listener) {
+    listener.call(HashLocation, change);
+  });
+}
+
 function ensureSlash() {
-  var path = getHashPath();
+  var path = HashLocation.getCurrentPath();
 
   if (path.charAt(0) === '/')
     return true;
@@ -25,24 +29,6 @@ function ensureSlash() {
 
   return false;
 }
-
-var _changeListeners = [];
-
-function notifyChange(type) {
-  if (type === LocationActions.PUSH)
-    History.length += 1;
-
-  var change = {
-    path: getHashPath(),
-    type: type
-  };
-
-  _changeListeners.forEach(function (listener) {
-    listener(change);
-  });
-}
-
-var _isListening = false;
 
 function onHashChange() {
   if (ensureSlash()) {
@@ -60,8 +46,8 @@ function onHashChange() {
  */
 var HashLocation = {
 
-  addChangeListener: function (listener) {
-    _changeListeners.push(listener);
+  addChangeListener(listener) {
+    _listeners.push(listener);
 
     // Do this BEFORE listening for hashchange.
     ensureSlash();
@@ -77,12 +63,12 @@ var HashLocation = {
     }
   },
 
-  removeChangeListener: function(listener) {
-    _changeListeners = _changeListeners.filter(function (l) {
+  removeChangeListener(listener) {
+    _listeners = _listeners.filter(function (l) {
       return l !== listener;
     });
 
-    if (_changeListeners.length === 0) {
+    if (_listeners.length === 0) {
       if (window.removeEventListener) {
         window.removeEventListener('hashchange', onHashChange, false);
       } else {
@@ -93,26 +79,32 @@ var HashLocation = {
     }
   },
 
-  push: function (path) {
+  push(path) {
     _actionType = LocationActions.PUSH;
-    window.location.hash = encodeURI(path);
+    window.location.hash = path;
   },
 
-  replace: function (path) {
+  replace(path) {
     _actionType = LocationActions.REPLACE;
     window.location.replace(
-      window.location.pathname + window.location.search + '#' + encodeURI(path)
+      window.location.pathname + window.location.search + '#' + path
     );
   },
 
-  pop: function () {
+  pop() {
     _actionType = LocationActions.POP;
     History.back();
   },
 
-  getCurrentPath: getHashPath,
+  getCurrentPath() {
+    return decodeURI(
+      // We can't use window.location.hash here because it's not
+      // consistent across browsers - Firefox will pre-decode it!
+      window.location.href.split('#')[1] || ''
+    );
+  },
 
-  toString: function () {
+  toString() {
     return '<HashLocation>';
   }
 
